# openapi.yaml
openapi: 3.0.4
info:
  description: API gateway for dataservices providing data access and management for
    IKlik services.
  title: IKlik Backend Services
  version: 1.0.0
servers:
- url: https://localhost:8000/api/v1
...
paths:
  /experiments:
    get:
      description: List experiments for a given user.
      operationId: getExperiments
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
          description: Successful operation
        '400':
          description: Invalid request
        '422':
          description: Validation exception
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: List experiments.
      tags:
      - experiments
    post:
      description: Create a new experiment under a given user.
      operationId: createExperiment
      requestBody:
        $ref: '#/components/requestBodies/Experiment'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Experiment'
          description: Successful operation
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new experiment.
      tags:
      - experiments
...
  /experiments/{experimentId}/results:
    get:
      description: Get list of files to download for results.
      operationId: getExperimentResults
      parameters:
      - description: ID to uniquely identify an experiment.
        explode: false
        in: path
        name: experimentId
        required: true
        schema:
          type: string
        style: simple
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid status value
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Get results for an experiment.
      tags:
      - experiments
      - results
    post:
      description: This can only be done by the logged in user.
      operationId: uploadResults
      parameters:
      - description: ID to uniquely identify an experiment.
        explode: false
        in: path
        name: experimentId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadResults'
        description: Results to upload along with metadata
      responses:
        '200':
          description: Successful operation
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Upload results for an experiment.
      tags:
      - experiments
      - results
  /users:
    post:
      description: This can only be done by the superuser.
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful operation
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a user.
      tags:
      - users
  /users/{username}:
    delete:
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - description: The username that needs to be deleted
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        '200':
          description: User deleted
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete user resource.
      tags:
      - users
    get:
      description: Get uses detail based on username.
      operationId: getUserByName
      parameters:
      - description: The name that needs to be fetched. Use users1 for testing
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successful operation
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Get user by username.
      tags:
      - users
    put:
      description: This can only be done by the logged in user or superuser.
      operationId: updateUser
      parameters:
      - description: name that need to be deleted
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad request
        '404':
          description: User not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Update user resource.
      tags:
      - users
  /infrastructure/videos:
    get:
      description: Fetch a list of all available videos.
      operationId: getVideos
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Video'
                type: array
          description: A list of videos
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Retrieve videos list
      tags:
      - videos
    post:
      description: Upload a new video to the infrastructure portal (Super User access
        required).
      operationId: createVideo
      requestBody:
        $ref: '#/components/requestBodies/Video'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
          description: Video created successfully
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid video upload data
        '415':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unsupported media type for video upload.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Video file validation error.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create video
      tags:
      - videos
  /infrastructure/videos/{id}:
    delete:
      description: Delete a specific video by ID (Super User access required).
      operationId: deleteVideo
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
          description: Video deleted
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Video not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Delete video
      tags:
      - videos
    get:
      description: Fetch a specific video by ID.
      operationId: getVideo
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
          description: Video details
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Video not found
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Retrieve video
      tags:
      - videos
...
components:
  requestBodies:
    Experiment:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExperimentInput'
      description: Experiment object that needs to be added to the store
    Video:
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/schemas/UploadVideo'
      description: Video object that needs to be added to the store
...
    User:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserInput'
      description: User object creation and update in store
  schemas:
    Experiment:
      example:
        networkConditions:
          delay: 50ms
          packetLoss: 2%
        createdAt: 2024-04-01 10:00:00+00:00
        metricsRequested:
        - PSNR
        - PSNR
        encodingParameters:
          codec: h264
          bitrate: 5000kbps
          resolution: 1920x1080
        description: An experiment testing video encodings.
        videoSources:
        - video1.mp4
        - video1.mp4
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        experimentName: My Test Experiment
        status: pending
      properties:
        id:
          format: uuid
          title: id
          type: string
        experimentName:
          example: My Test Experiment
          title: experimentName
          type: string
        description:
          example: An experiment testing video encodings.
          title: description
          type: string
        videoSources:
          items:
            example: video1.mp4
            type: string
          title: videoSources
          type: array
        encodingParameters:
          $ref: '#/components/schemas/EncodingParameters'
        networkConditions:
          $ref: '#/components/schemas/NetworkConditions'
        metricsRequested:
          items:
            example: PSNR
            type: string
          title: metricsRequested
          type: array
        status:
          example: pending
          title: status
          type: string
        createdAt:
          example: 2024-04-01 10:00:00+00:00
          format: date-time
          title: createdAt
          type: string
      required:
      - encodingParameters
      - experimentName
      - metricsRequested
      - networkConditions
      - videoSources
      title: Experiment
      type: object
    ExperimentInput:
      example:
        networkConditions:
          delay: 100ms
          packetLoss: 1%
        metricsRequested:
        - PSNR
        - SSIM
        encodingParameters:
          codec: H.264
          bitrate: 5000kbps
          resolution: 1920x1080
        description: Experiment description text here
        videoSources:
        - video1.mp4
        - video2.mp4
        experimentName: Video Encoding Test
      properties:
        experimentName:
          example: Video Encoding Test
          title: experimentName
          type: string
        description:
          example: Experiment description text here
          title: description
          type: string
        videoSources:
          example:
          - video1.mp4
          - video2.mp4
          items:
            type: string
          title: videoSources
          type: array
        encodingParameters:
          $ref: '#/components/schemas/ExperimentInput'
        networkConditions:
          $ref: '#/components/schemas/ExperimentInput'
        metricsRequested:
          example:
          - PSNR
          - SSIM
          items:
            type: string
          title: metricsRequested
          type: array
      required:
      - encodingParameters
      - experimentName
      - metricsRequested
      - networkConditions
      - videoSources
      title: ExperimentInput
      type: object
    User:
      example:
        firstName: John
        lastName: James
        password: '12345'
        id: 10
        email: john@email.com
        username: theUser
      properties:
        id:
          example: 10
          format: int64
          title: id
          type: integer
        username:
          example: theUser
          title: username
          type: string
        firstName:
          example: John
          title: firstName
          type: string
        lastName:
          example: James
          title: lastName
          type: string
        email:
          example: john@email.com
          title: email
          type: string
        password:
          example: '12345'
          title: password
          type: string
        role:
          title: role
          type: string
          enum:
          - user
          - admin
          - super_admin
          default: user
      title: User
      type: object
    UserInput:
      example:
        firstName: John
        lastName: James
        password: '12345'
        email: john@email.com
        username: theUser
      properties:
        username:
          example: theUser
          title: username
          type: string
        firstName:
          example: John
          title: firstName
          type: string
        lastName:
          example: James
          title: lastName
          type: string
        email:
          example: john@email.com
          title: email
          type: string
        password:
          example: '12345'
          title: password
          type: string
        role:
          title: role
          type: string
          enum:
          - user
          - admin
          - super_admin
          default: user
      title: UserInput
      type: object
    Error:
      example:
        code: code
        message: message
      properties:
        code:
          title: code
          type: string
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: Error
      type: object
    Video:
      example:
        frameRate: frameRate
        createdDate: 2000-01-23 04:56:07+00:00
        id: id
        video: ''
        resolution: resolution
      properties:
        id:
          title: id
          type: string
        video:
          format: binary
          title: video
          type: string
        frameRate:
          title: frameRate
          type: string
        resolution:
          title: resolution
          type: string
        createdDate:
          format: date-time
          title: createdDate
          type: string
      title: Video
      type: object
...
    UploadVideo:
      properties:
        video:
          format: binary
          type: string
        frameRate:
          type: string
        resolution:
          type: string
      type: object
...
    EncodingParameters:
      example:
        codec: h264
        bitrate: 5000kbps
        resolution: 1920x1080
      properties:
        codec:
          example: h264
          title: codec
          type: string
        bitrate:
          example: 5000kbps
          title: bitrate
          type: string
        resolution:
          example: 1920x1080
          title: resolution
          type: string
      title: EncodingParameters
      type: object
    NetworkConditions:
      example:
        delay: 50ms
        packetLoss: 2%
      properties:
        packetLoss:
          example: 2%
          title: packetLoss
          type: string
        delay:
          example: 50ms
          title: delay
          type: string
      title: NetworkConditions
      type: object
    UploadResults:
      properties:
        filename:
          items:
            format: binary
            type: string
          type: array
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []